move gops code to ps.go
